requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.102.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.6
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
